+++ src/hhtznr/josm/plugins/elevation/ElevationPreferences.java
@@ -331,9 +331,15 @@ public class ElevationPreferences {
      */
     public static final int INCR_HILLSHADE_AZIMUTH = 1;
 
-    public static final int DEFAULT_LOWER_CUTOFF_ELEVATION = -999;
+    /**
+     * Default value of the lower cutoff elevation: {@code -999}.
+     */
+    public static final short DEFAULT_LOWER_CUTOFF_ELEVATION = -999;
 
-    public static final int DEFAULT_UPPER_CUTOFF_ELEVATION = 9999;
+    /**
+     * Default of the upper cutoff elevation: {@code 9999}.
+     */
+    public static final short DEFAULT_UPPER_CUTOFF_ELEVATION = 9999;
 
     /**
      * Property key for enabling or disabling rendering of the elevation data point
